+++ drivers/imu/adis.c
@@ -42,6 +42,7 @@
 /******************************************************************************/
 
 #include "adis.h"
+#include "adis_internals.h"
 #include "no_os_delay.h"
 #include "no_os_gpio.h"
 #include "no_os_util.h"
